// src/gql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** A date string, such as 2007-12-03 (YYYY-MM-DD), compliant with ISO 8601 standard for representation of dates using the Gregorian calendar. */
  Date: { input: unknown; output: unknown; }
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the date-timeformat outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representationof dates and times using the Gregorian calendar. */
  DateTime: { input: unknown; output: unknown; }
  Hex: { input: unknown; output: unknown; }
  /** Raw JSON value */
  Json: { input: unknown; output: unknown; }
  /** The Long scalar type represents non-fractional signed whole numeric values. Long can represent values between -(2^63) and 2^63 - 1. */
  Long: { input: unknown; output: unknown; }
  RGBAHue: { input: unknown; output: unknown; }
  RGBATransparency: { input: unknown; output: unknown; }
  /** Slate-compatible RichText AST */
  RichTextAST: { input: unknown; output: unknown; }
};
// ...
export type Category = Node & {
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  description?: Maybe<Scalars['String']['output']>;
  /** Get the document in other stages */
  documentInStages: Array<Category>;
  /** List of Category versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  /** System Locale field */
  locale: Locale;
  /** Get the other localizations for this document */
  localizations: Array<Category>;
  name: Scalars['String']['output'];
  products: Array<Product>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  scheduledIn: Array<ScheduledOperation>;
  slug: Scalars['String']['output'];
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
};
// ...
export type Collection = Node & {
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  description?: Maybe<Scalars['String']['output']>;
  /** Get the document in other stages */
  documentInStages: Array<Collection>;
  /** List of Collection versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  image: Asset;
  /** System Locale field */
  locale: Locale;
  /** Get the other localizations for this document */
  localizations: Array<Collection>;
  name: Scalars['String']['output'];
  products: Array<Product>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  scheduledIn: Array<ScheduledOperation>;
  slug: Scalars['String']['output'];
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
};
// ...
export type Locale =
  /** System locale */
  | 'en';
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars['ID']['output'];
  /** The Stage of an object */
  stage: Stage;
};
// ...
export type Product = Node & {
  averageRating: Scalars['Int']['output'];
  categories: Array<Category>;
  collections: Array<Collection>;
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  description: Scalars['String']['output'];
  /** Get the document in other stages */
  documentInStages: Array<Product>;
  /** List of Product versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  images: Array<Asset>;
  /** System Locale field */
  locale: Locale;
  /** Get the other localizations for this document */
  localizations: Array<Product>;
  name: Scalars['String']['output'];
  orderItems: Array<OrderItem>;
  price: Scalars['Int']['output'];
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  reviews: Array<Review>;
  scheduledIn: Array<ScheduledOperation>;
  slug: Scalars['String']['output'];
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
  variants: Array<ProductVariants>;
};
// ...
export type ProductOrderByInput =
  | 'averageRating_ASC'
  | 'averageRating_DESC'
  | 'createdAt_ASC'
  | 'createdAt_DESC'
  | 'description_ASC'
  | 'description_DESC'
  | 'id_ASC'
  | 'id_DESC'
  | 'name_ASC'
  | 'name_DESC'
  | 'price_ASC'
  | 'price_DESC'
  | 'publishedAt_ASC'
  | 'publishedAt_DESC'
  | 'slug_ASC'
  | 'slug_DESC'
  | 'updatedAt_ASC'
  | 'updatedAt_DESC';
// ...
export type Review = Node & {
  content: Scalars['String']['output'];
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  /** Get the document in other stages */
  documentInStages: Array<Review>;
  email: Scalars['String']['output'];
  headline: Scalars['String']['output'];
  /** List of Review versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  product?: Maybe<Product>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  rating: Scalars['Int']['output'];
  scheduledIn: Array<ScheduledOperation>;
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
};
// ...
export type ProductGetByIdQueryVariables = Exact<{
  id: Scalars['ID']['input'];
}>;


export type ProductGetByIdQuery = { product?: { id: string, name: string, description: string, price: number, averageRating: number, categories: Array<{ name: string }>, images: Array<{ url: string }>, reviews: Array<{ id: string, name: string, headline: string, email: string, content: string, rating: number, stage: Stage }> } | null };
// ...
export type ProductsGetByCategorySlugQueryVariables = Exact<{
  slug: Scalars['String']['input'];
  first?: InputMaybe<Scalars['Int']['input']>;
  skip?: InputMaybe<Scalars['Int']['input']>;
}>;


export type ProductsGetByCategorySlugQuery = { categories: Array<{ name: string, description?: string | null, products: Array<{ id: string, name: string, description: string, price: number, averageRating: number, categories: Array<{ name: string }>, images: Array<{ url: string }>, reviews: Array<{ id: string, name: string, headline: string, email: string, content: string, rating: number, stage: Stage }> }> }>, productsConnection: { pageInfo: { pageSize?: number | null } } };

export type ProductsGetByCollectionSlugQueryVariables = Exact<{
  slug: Scalars['String']['input'];
}>;


export type ProductsGetByCollectionSlugQuery = { collections: Array<{ name: string, description?: string | null, products: Array<{ id: string, name: string, description: string, price: number, averageRating: number, categories: Array<{ name: string }>, images: Array<{ url: string }>, reviews: Array<{ id: string, name: string, headline: string, email: string, content: string, rating: number, stage: Stage }> }> }> };

export type ProductsGetListQueryVariables = Exact<{
  first?: InputMaybe<Scalars['Int']['input']>;
  skip?: InputMaybe<Scalars['Int']['input']>;
  query?: InputMaybe<Scalars['String']['input']>;
  orderBy?: InputMaybe<ProductOrderByInput>;
}>;


export type ProductsGetListQuery = { products: Array<{ id: string, name: string, description: string, price: number, averageRating: number, categories: Array<{ name: string }>, images: Array<{ url: string }>, reviews: Array<{ id: string, name: string, headline: string, email: string, content: string, rating: number, stage: Stage }> }>, productsConnection: { pageInfo: { pageSize?: number | null } } };
// ...
export class TypedDocumentString<TResult, TVariables>
  extends String
  implements DocumentTypeDecoration<TResult, TVariables>
{
  __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType'];

  constructor(private value: string, public __meta__?: Record<string, any>) {
    super(value);
  }

  toString(): string & DocumentTypeDecoration<TResult, TVariables> {
    return this.value;
  }
}
// ...
export const ProductGetByIdDocument = new TypedDocumentString(`
    query ProductGetById($id: ID!) {
  product(where: {id: $id}) {
    ...ProductItem
  }
}
    fragment ProductItem on Product {
  id
  name
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    url
  }
  price
  reviews {
    ...ReviewsItem
  }
  averageRating
}
fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}`) as unknown as TypedDocumentString<ProductGetByIdQuery, ProductGetByIdQueryVariables>;
// ...
export const ProductsGetByCategorySlugDocument = new TypedDocumentString(`
    query ProductsGetByCategorySlug($slug: String!, $first: Int, $skip: Int) {
  categories(where: {slug: $slug}) {
    name
    description
    products(first: $first, skip: $skip) {
      ...ProductItem
    }
  }
  productsConnection(where: {categories_some: {slug: $slug}}) {
    pageInfo {
      pageSize
    }
  }
}
    fragment ProductItem on Product {
  id
  name
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    url
  }
  price
  reviews {
    ...ReviewsItem
  }
  averageRating
}
fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}`) as unknown as TypedDocumentString<ProductsGetByCategorySlugQuery, ProductsGetByCategorySlugQueryVariables>;
export const ProductsGetByCollectionSlugDocument = new TypedDocumentString(`
    query ProductsGetByCollectionSlug($slug: String!) {
  collections(where: {slug: $slug}) {
    name
    description
    products {
      ...ProductItem
    }
  }
}
    fragment ProductItem on Product {
  id
  name
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    url
  }
  price
  reviews {
    ...ReviewsItem
  }
  averageRating
}
fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}`) as unknown as TypedDocumentString<ProductsGetByCollectionSlugQuery, ProductsGetByCollectionSlugQueryVariables>;
export const ProductsGetListDocument = new TypedDocumentString(`
    query ProductsGetList($first: Int, $skip: Int, $query: String, $orderBy: ProductOrderByInput) {
  products(
    where: {name_contains: $query}
    first: $first
    skip: $skip
    orderBy: $orderBy
  ) {
    ...ProductItem
  }
  productsConnection(where: {name_contains: $query}) {
    pageInfo {
      pageSize
    }
  }
}
    fragment ProductItem on Product {
  id
  name
  description
  categories(first: 1) {
    name
  }
  images(first: 1) {
    url
  }
  price
  reviews {
    ...ReviewsItem
  }
  averageRating
}
fragment ReviewsItem on Review {
  id
  name
  headline
  email
  content
  rating
  stage
}`) as unknown as TypedDocumentString<ProductsGetListQuery, ProductsGetListQueryVariables>;